+++ example/minimal/lib/main.dart
@@ -166,7 +166,7 @@ final class AppDependencies implements ScopeDependencies {
   }
 
   @override
-  void unmount() {}
+  void onUnmount() {}
 
   @override
   Future<void> dispose() async {}